Rename module Route to Map, rename some other variables
